Update: remove utils
