--- bcs-services/bcs-webconsole/console/utils/utils.go
@@ -1,48 +1,0 @@
-/*
- * Tencent is pleased to support the open source community by making Blueking Container Service available.
- * Copyright (C) 2019 THL A29 Limited, a Tencent company. All rights reserved.
- * Licensed under the MIT License (the "License"); you may not use this file except
- * in compliance with the License. You may obtain a copy of the License at
- * http://opensource.org/licenses/MIT
- * Unless required by applicable law or agreed to in writing, software distributed under
- * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
- * either express or implied. See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-
-package utils
-
-import (
-	"net/http"
-	"os"
-
-	"github.com/Tencent/bk-bcs/bcs-services/bcs-webconsole/console/config"
-	"github.com/Tencent/bk-bcs/bcs-services/bcs-webconsole/console/types"
-
-	"github.com/gin-gonic/gin"
-	"github.com/google/uuid"
-	"github.com/pkg/errors"
-)
-
-// APIError 简易的错误返回
-func APIError(c *gin.Context, msg string) {
-	data := types.APIResponse{
-		Code:      types.ApiErrorCode,
-		Message:   msg,
-		RequestID: uuid.New().String(),
-	}
-
-	c.AbortWithStatusJSON(http.StatusOK, data)
-}
-
-func GetUsername(c *gin.Context) (string, error) {
-	// DEV环境
-	if config.G.Base.RunEnv == config.DevEnv {
-		username := os.Getenv("WEBCONSOLE_USERNAME")
-		if username != "" {
-			return username, nil
-		}
-	}
-	return "", errors.New("username 不能为空")
-}
